refactor(tasks): move locomotion command helpers [skip ci]
