+++ src/unilab/envs/locomotion/common/__init__.py
@@ -6,13 +6,6 @@ from .base import (
     PdControlConfig,
     Sensor,
 )
-from .commands import (
-    Commands,
-    apply_heading_yaw_feedback,
-    sample_heading_commands,
-    sample_velocity_commands,
-    zero_small_xy_commands,
-)
 from .domain_rand import DomainRandConfig
 from .dr_provider import LocomotionDRProvider
 from .height_scan import (
@@ -24,7 +17,6 @@ from .rewards import RewardContext
 
 __all__ = [
     "BaseNoiseConfig",
-    "Commands",
     "ControlConfigBase",
     "DEFAULT_SCAN_POINTS_X",
     "DEFAULT_SCAN_POINTS_Y",
@@ -36,8 +28,4 @@ __all__ = [
     "PdControlConfig",
     "RewardContext",
     "Sensor",
-    "apply_heading_yaw_feedback",
-    "sample_heading_commands",
-    "sample_velocity_commands",
-    "zero_small_xy_commands",
 ]
