+++ src/unilab/tasks/locomotion/common/__init__.py
@@ -1,1 +1,17 @@
 """Shared task-specific locomotion components."""
+
+from .commands import (
+    Commands,
+    apply_heading_yaw_feedback,
+    sample_heading_commands,
+    sample_velocity_commands,
+    zero_small_xy_commands,
+)
+
+__all__ = [
+    "Commands",
+    "apply_heading_yaw_feedback",
+    "sample_heading_commands",
+    "sample_velocity_commands",
+    "zero_small_xy_commands",
+]
