+++ src/unilab/tasks/locomotion/common/commands.py
@@ -1,3 +1,5 @@
+"""Shared command helpers for locomotion tasks."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass, field
