+++ src/unilab/tasks/locomotion/go2/rough.py
@@ -13,11 +13,6 @@ from unilab.dr import DomainRandomizationManager, ResetPlan
 from unilab.dr.dr_utils import build_common_reset_randomization, zero_actions
 from unilab.dtype_config import get_global_dtype
 from unilab.envs.locomotion.common import rewards
-from unilab.envs.locomotion.common.commands import (
-    apply_heading_yaw_feedback,
-    sample_heading_commands,
-    zero_small_xy_commands,
-)
 from unilab.envs.locomotion.common.height_scan import (
     DEFAULT_SCAN_POINTS_X,
     DEFAULT_SCAN_POINTS_Y,
@@ -29,6 +24,11 @@ from unilab.envs.locomotion.common.height_scan import (
     terrain_out_of_bounds,
 )
 from unilab.envs.locomotion.common.rewards import RewardContext
+from unilab.tasks.locomotion.common.commands import (
+    apply_heading_yaw_feedback,
+    sample_heading_commands,
+    zero_small_xy_commands,
+)
 from unilab.tasks.locomotion.go2.base import ControlConfig
 from unilab.tasks.locomotion.go2.joystick import (
     Commands,
